Random port assignment to a service
