+++ .gitignore
@@ -9,6 +9,7 @@
 .project
 .settings
 
+code/packaging/app/keys.properties
 dist
 docs/build
 resources/content/db/mysql/cattle_dump*.sql
@@ -29,3 +30,4 @@ tests/integration/MANIFEST
 *.pyo
 *.swp
 *~
+
