+++ code/iaas/model/src/main/java/io/cattle/platform/core/constants/AccountConstants.java
@@ -11,6 +11,7 @@ public class AccountConstants {
     public static final String USER_KIND = "user";
     public static final String ADMIN_KIND = "admin";
     public static final String PROJECT_KIND = "project";
+    public static final String FIELD_PORT_RANGE = "servicesPortRange";
 
     public static final String ACCOUNT_ID = ConstantsUtils.property(Account.class, "accountId");
     public static final String SYSTEM_UUID = "system";
