+++ code/iaas/resource-pool/src/main/java/io/cattle/platform/resource/pool/util/ResourcePoolConstants.java
@@ -6,5 +6,5 @@ public class ResourcePoolConstants {
     public static final String MAC = "mac";
     public static final String LINK_PORT = "linkPort";
     public static final String HOST_PORT = "hostPort";
-
+    public static final String ENVIRONMENT_PORT = "environmentPort";
 }
