+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/process/ServiceCreate.java
@@ -33,6 +33,7 @@ public class ServiceCreate extends AbstractObjectProcessHandler {
     public HandlerResult handle(ProcessState state, ProcessInstance process) {
         Service service = (Service) state.getResource();
         sdService.setVIP(service);
+        sdService.setPorts(service);
         return null;
     }
 }
