+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/process/ServiceRemove.java
@@ -43,6 +43,8 @@ public class ServiceRemove extends AbstractObjectProcessHandler {
 
         sdService.releaseVip(service);
 
+        sdService.releasePorts(service);
+
         return null;
     }
 }
