+++ code/packaging/app-config/src/main/resources/META-INF/cattle/system-services/pool-defaults.properties
@@ -7,3 +7,5 @@ link.internal.port.end=39999
 
 host.port.start=20000
 host.port.end=39999
+
+environment.services.port.range=49153-65535
